Cloud Function_ Tabla Hechos particionada/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery>=1.27.0